fix: Assert `--remove-prefix` option is ?string
The rector run in CI runs on PHP 8.1, which means it also installs
symfony 5.4. In symfony/console 5.4, `getOption` has no strict return
type so Rector ignores use of it.

Newer symfony explicitly types this as `:mixed`, so running Rector on a
later php version (e.g. locally) will by default add a (string) cast to
the `explode` statement. However, in practice this is unnecessary
because symfony/console guarantees that an option with
`InputOption::VALUE_REQUIRED` will give a value of `?string`.