refactor: Improve type-safety of preg_replace operations
`preg_replace` can return null in some cases (e.g. if the pattern is
valid but it hits the runtime backtrack_limit). We were not always
handling this - sometimes casting back to empty string, sometimes just
assuming the result would be a string.

This can also happen if there is an issue with the regex itself (e.g.
the pattern is invalid) and the runtime PHP error is not captured.

It can also theoretically return an array if `$subject` was an array,
but we never use it like this.

This commit introduces a thin wrapper function to enforce that the
preg_replace call was successful, to reduce the need for assertions
elsewhere in the code as we move towards stricter typing.